chore: [hub] rename folder
