+++ hub/src/main/java/com/sparta/hub/application/initializer/HubAndRouteInitializerService.java
@@ -4,8 +4,8 @@ import com.sparta.hub.domain.map.model.Coordinates;
 import com.sparta.hub.domain.map.service.MapService;
 import com.sparta.hub.domain.hub.model.Hub;
 import com.sparta.hub.domain.hub.service.HubServiceImpl;
-import com.sparta.hub.domain.hubRoute.model.HubRoute;
-import com.sparta.hub.domain.hubRoute.service.HubRouteService;
+import com.sparta.hub.domain.route.model.HubRoute;
+import com.sparta.hub.domain.route.service.HubRouteService;
 import com.sparta.hub.infrastructure.persistence.HubRepository;
 import com.sparta.hub.infrastructure.persistence.HubRouteRepository;
 import lombok.RequiredArgsConstructor;
