+++ hub/src/main/java/com/sparta/hub/domain/route/dto/HubRouteResponseDto.java
@@ -1,7 +1,7 @@
-package com.sparta.hub.domain.hubRoute.dto;
+package com.sparta.hub.domain.route.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.sparta.hub.domain.hubRoute.model.HubRoute;
+import com.sparta.hub.domain.route.model.HubRoute;
 import java.util.UUID;
 import lombok.AccessLevel;
 import lombok.Builder;
