+++ hub/src/main/java/com/sparta/hub/domain/route/service/cache/HubRouteCacheService.java
@@ -1,8 +1,8 @@
-package com.sparta.hub.domain.hubRoute.service.cache;
+package com.sparta.hub.domain.route.service.cache;
 
-import com.sparta.hub.domain.hubRoute.dto.HubRouteDto;
-import com.sparta.hub.domain.hubRoute.dto.HubRouteResponseDto;
-import com.sparta.hub.domain.hubRoute.model.HubRoute;
+import com.sparta.hub.domain.route.dto.HubRouteDto;
+import com.sparta.hub.domain.route.dto.HubRouteResponseDto;
+import com.sparta.hub.domain.route.model.HubRoute;
 import com.sparta.hub.infrastructure.persistence.HubRouteRepository;
 import jakarta.persistence.EntityNotFoundException;
 import java.util.List;
